modul_6/DB_Postgres/pyproject.toml
[tool.poetry]
name = "db-postgres"
version = "0.1.0"
description = ""
authors = ["Aleksandr <aleksandr.voitushenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
faker = "^24.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
